remove jwt warning
